refactor: add new folder structure
