+++ src/index.tsx
@@ -1,7 +1,7 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+import App from './app/App';
 import reportWebVitals from './reportWebVitals';
 
 import './styles/globals.css';
